+++ lib/cacheConnection.js
@@ -12,16 +12,13 @@ function applyTimeoutIfNeeded(pending, timeout) {
 
 function isCacheReady({ cache, timeout }, callback) {
   applyTimeoutIfNeeded(cache.start(), timeout)
-    .then(() => {
-      callback();
-    })
-    .catch((err) => {
-      callback(err);
-    });
+    .then(() => callback())
+    .catch(callback);
 }
 
 async function startCacheConnection({cache, timeout, circuitBreaker }) {
   if (!circuitBreaker) return applyTimeoutIfNeeded(cache.start(), timeout);
+
   return new Promise((resolve, reject) => {
     circuitBreaker.run({cache, timeout }, (err) => {
       if (err) return reject(err);
