+++ lib/maxAge.js
@@ -10,8 +10,8 @@ const toResponse = require('./toResponse');
 const isCachable = require('./isCachable');
 const events = require('./events');
 const directives = require('./directives');
-const startCacheConnection = require('./startCacheConnection').startCacheConnection;
-const isCacheReady = require('./startCacheConnection').isCacheReady;
+const startCacheConnection = require('./cacheConnection').startCacheConnection;
+const isCacheReady = require('./cacheConnection').isCacheReady;
 const MAX_AGE = 'max-age';
 const CACHE_CONTROL = 'cache-control';
 const SEGMENT = 'body';
