+++ lib/staleIfError.js
@@ -9,8 +9,8 @@ const toResponse = require('./toResponse');
 const isCachable = require('./isCachable');
 const events = require('./events');
 const directives = require('./directives');
-const startCacheConnection = require('./startCacheConnection').startCacheConnection;
-const isCacheReady = require('./startCacheConnection').isCacheReady;
+const startCacheConnection = require('./cacheConnection').startCacheConnection;
+const isCacheReady = require('./cacheConnection').isCacheReady;
 
 const STALE_IF_ERROR = 'stale-if-error';
 const MAX_AGE = 'max-age';
